Initial commit
This contains most of the preliminary work.
This version is not ready to be deployed yet, because i ommited
character information and picture from the commit.

My next plan is to make people able to load the character rules, as a
json and the character picture as a link. Character will be saved in
local storage, or as a json file.

This character sheet is still missing diseases and other debuff that can
take a dice space. It should also be able to roll attacks.